more devices +  some tweaks  to latch, types, device

+++ devices/types.hh
@@ -1,6 +1,6 @@
 #ifndef TYPES_HH
 #define TYPES_HH
 
-typedef double Port;
+typedef long long Port;
 
 #endif
